# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.FormAnalytics:
    properties:
      average_score:
        type: number
      highest_score:
        type: number
      lowest_score:
        type: number
      question_breakdown:
        additionalProperties:
          $ref: '#/definitions/domain.QuestionAnalytics'
        type: object
      total_responses:
        type: integer
    type: object
...
  domain.QuestionAnalytics:
    properties:
      accuracy_rate:
        type: number
      correct_count:
        type: integer
      option_counts:
        additionalProperties:
          type: integer
        description: Option ID -> count
        type: object
      question_id:
        type: string
      question_text:
        type: string
      total_answered:
        type: integer
    type: object
...
  dto.AnswerDetailDTO:
    properties:
      answer_text:
        type: string
      id:
        type: string
      is_correct:
        type: boolean
      is_flagged:
        type: boolean
      match_pair_json:
        type: string
      points_earned:
        type: number
      question_id:
        type: string
      question_text:
        type: string
      score_given:
        type: number
      selected_option_id:
        type: string
      selected_option_text:
        type: string
    type: object
...
  dto.AutosaveAnswerRequest:
    properties:
      answer_text:
        type: string
      is_flagged:
        type: boolean
      match_pairs:
        items:
          $ref: '#/definitions/dto.MatchPairItem'
        type: array
      question_id:
        type: string
      selected_option_id:
        type: string
    required:
    - question_id
    type: object
  dto.AutosaveResponse:
    properties:
      is_flagged:
        type: boolean
      message:
        type: string
      question_id:
        type: string
      saved_at:
        type: string
      success:
        type: boolean
    type: object
...
  dto.GradeResponseRequest:
    properties:
      total_score:
        minimum: 0
        type: number
    type: object
...
  dto.MatchPairItem:
    properties:
      match_key:
        type: string
      match_target_text:
        type: string
    type: object
...
  dto.ResponseDetailDTO:
    properties:
      answers:
        items:
          $ref: '#/definitions/dto.AnswerDetailDTO'
        type: array
      device_platform:
        type: string
      form_id:
        type: string
      id:
        type: string
      is_auto_submitted:
        type: boolean
      respondent_email:
        type: string
      started_at:
        type: string
      status:
        type: string
      submitted_at:
        type: string
      total_score:
        type: number
    type: object
  dto.RestartStudentSessionRequest:
    properties:
      warning_message:
        type: string
    required:
    - warning_message
    type: object
  dto.SessionQuestionItemDTO:
    properties:
      answer_text:
        type: string
      is_answered:
        type: boolean
      is_flagged:
        type: boolean
      match_pairs:
        items:
          $ref: '#/definitions/dto.MatchPairItem'
        type: array
      order_index:
        type: integer
      question_id:
        type: string
      selected_option_id:
        type: string
    type: object
  dto.SessionStateDTO:
    properties:
      current_question_index:
        type: integer
      duration_minutes:
        type: integer
      form_id:
        type: string
      is_warning_acknowledged:
        type: boolean
      questions:
        items:
          $ref: '#/definitions/dto.SessionQuestionItemDTO'
        type: array
      response_id:
        type: string
      started_at:
        type: string
      status:
        type: string
      warning_message:
        type: string
    type: object
  dto.SubmitAnswerDetail:
    properties:
      answer_text:
        type: string
      is_flagged:
        type: boolean
      match_pairs:
        items:
          $ref: '#/definitions/dto.MatchPairItem'
        type: array
      question_id:
        type: string
      selected_option_id:
        type: string
    required:
    - question_id
    type: object
  dto.SubmitFormRequest:
    properties:
      answers:
        items:
          $ref: '#/definitions/dto.SubmitAnswerDetail'
        type: array
      device_platform:
        type: string
      is_auto_submitted:
        type: boolean
      passcode:
        type: string
      respondent_email:
        type: string
      response_id:
        type: string
    required:
    - answers
    - respondent_email
    type: object
  dto.SubmitResponseResult:
    properties:
      is_auto_submitted:
        type: boolean
      message:
        type: string
      response_id:
        type: string
      submitted_at:
        type: string
      total_score:
        type: number
    type: object
...
  dto.TelemetryEventRequest:
    properties:
      current_question_index:
        type: integer
      event_message:
        type: string
      event_type:
        type: string
      metadata:
        type: string
    required:
    - event_type
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@hidocs.id
    name: HiDocs Support
  description: High-performance backend API for HiDocs Form & Exam Maker application
    using Golang and Domain-Driven Design (DDD).
  termsOfService: http://swagger.io/terms/
  title: HiDocs Backend API (Form & Exam Maker)
  version: "1.0"
paths:
...
  /api/v1/forms/{form_id}/analytics:
    get:
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.FormAnalytics'
              type: object
      security:
      - BearerAuth: []
      summary: Get real-time analytics for charts (Pie Chart score distribution, Question
        accuracy bar charts)
      tags:
      - Analytics
  /api/v1/forms/{form_id}/export:
    get:
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: 'Export format (''xlsx'' or ''csv'', default: ''xlsx'')'
        in: query
        name: format
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Export form responses to Excel or CSV
      tags:
      - Responses
...
  /api/v1/forms/{form_id}/responses:
    get:
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ResponseDetailDTO'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get all responses for a form
      tags:
      - Responses
  /api/v1/forms/{form_id}/responses/{response_id}/restart:
    post:
      consumes:
      - application/json
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Response ID
        in: path
        name: response_id
        required: true
        type: string
      - description: Restart Reason Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RestartStudentSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Creator restarts student exam attempt due to suspicious cheating behavior
      tags:
      - Live Monitoring
...
  /api/v1/forms/{form_id}/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Submit Form Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitFormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SubmitResponseResult'
              type: object
      summary: Submit form / exam answers
      tags:
      - Responses
...
  /api/v1/public/responses/{response_id}/acknowledge-warning:
    post:
      parameters:
      - description: Response Session ID
        in: path
        name: response_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Student acknowledges proctor warning after exam restart
      tags:
      - Public
  /api/v1/public/responses/{response_id}/autosave:
    post:
      consumes:
      - application/json
      parameters:
      - description: Response Session ID
        in: path
        name: response_id
        required: true
        type: string
      - description: Autosave Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AutosaveAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AutosaveResponse'
              type: object
      summary: Autosave incremental student answer and flagged (ragu-ragu) state
      tags:
      - Public
  /api/v1/public/responses/{response_id}/session:
    get:
      parameters:
      - description: Response Session ID
        in: path
        name: response_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SessionStateDTO'
              type: object
      summary: Get active exam session state and questions status (answered / flagged)
        for navigation
      tags:
      - Public
  /api/v1/public/responses/{response_id}/telemetry:
    post:
      consumes:
      - application/json
      parameters:
      - description: Response Session ID
        in: path
        name: response_id
        required: true
        type: string
      - description: Telemetry Event Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TelemetryEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Send client telemetry events (tab switch, window blur, app background,
        screenshot, split screen)
      tags:
      - Public
...
  /api/v1/responses/{response_id}:
    get:
      parameters:
      - description: Response ID
        in: path
        name: response_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResponseDetailDTO'
              type: object
      security:
      - BearerAuth: []
      summary: Get individual response details
      tags:
      - Responses
  /api/v1/responses/{response_id}/grade:
    put:
      consumes:
      - application/json
      parameters:
      - description: Response ID
        in: path
        name: response_id
        required: true
        type: string
      - description: Grade Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GradeResponseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Adjust score/grade manually for a response
      tags:
      - Responses
  /api/v1/responses/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ResponseDetailDTO'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get current user's own form submissions
      tags:
      - Responses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"